Improve promises / add detect display changes
Litfaß now uses a strictly coherent promise chain to start and run. The
promise returned by the start method will resolve, as soon as all
browsers (or any of their pages) are closed. This way Litfaß will exit
and shut down the local server with it.

A new watchNumberOfDisplays property has been introduced to watch
for the number of displays connected. This way Litfaß will automatically
adapt if displays are attached / detached.